Move testes DetalharFluxoView para arquivo próprio #63
